Various changes
Problem: The code as it is right now could use a little cleanup, and
using turn curvature directly is not an ideal mechanism for steering.

Solution: Implement proper steering using the steering wheel angle
instead of turn curvature to determine the motion of the vehicle. Also
implemented simple damping for vehicle speed, and removed debug lines.